+++ src/test/java/space/wangjiang/summer/util/FileUtilTest.java
@@ -23,23 +23,4 @@ public class FileUtilTest extends BaseTest {
         EasyLogger.debug(FileUtil.getSystemTempDir());
     }
 
-    @Test
-    public void readTest() {
-        File file = new File("D:/test.txt");
-        String res = FileUtil.readFileToString(file, "GB2312");
-        EasyLogger.debug(res);
-    }
-
-
-    @Test
-    public void writeTest() throws IOException {
-        String fileName = StringUtil.getRandomString(10);
-        String data = "中文测试" + StringUtil.getRandomString(20);
-        File tempFile = new File(FileUtil.getSystemTempDir() + fileName);
-        FileUtil.write(tempFile, data, "UTF-8");
-        EasyLogger.debug(tempFile.getAbsolutePath());
-        String res = FileUtil.readFileToString(tempFile, "UTF-8");
-        Assert.assertEquals(data, res);
-    }
-
 }