+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # 更新日志
 
+## 0.1.2 (2018-04-17)
+* [x] 将Dialect相关类移到了dialect包
+* [x] 增加了ConnectionProvider，移除了对Druid的强制依赖
+* [x] IPlugin接口名修改为Plugin
+* [x] 文件上传库由cos改为commons-fileupload
+* [x] 重构了整个upload，移除了UploadRequestWrapper，改为UploadRequest，增加了上传策略、文件重命名策略等功能
+
 ## 0.1.1 (2018-04-09)
 * [x] 0.1.0 Maven配置错误导致jitpack构建错误